Rename migthy_types to mighty_types and delete second init file
