+++ mighty/mighty_agents/ppo.py
@@ -13,7 +13,7 @@ from mighty.mighty_models.ppo import PPOModel
 from mighty.mighty_replay import RolloutBatch
 from mighty.mighty_replay.mighty_rollout_buffer import MightyRolloutBuffer
 from mighty.mighty_update.ppo_update import PPOUpdate
-from mighty.mighty_utils.migthy_types import MIGHTYENV, TypeKwargs
+from mighty.mighty_utils.mighty_types import MIGHTYENV, TypeKwargs
 
 
 class MightyPPOAgent(MightyAgent):
