--- mighty/mighty_utils/__init__ .py
@@ -1,12 +1,0 @@
-from types import MIGHTYENV, TypeKwargs, retrieve_class
-
-from envs import make_mighty_env
-from update_utils import polyak_update
-
-__all__ = [
-    "MIGHTYENV",
-    "make_mighty_env",
-    "TypeKwargs",
-    "retrieve_class",
-    "polyak_update",
-]
