+++ mighty/mighty_utils/__init__.py
@@ -1,9 +1,11 @@
 from .envs import make_mighty_env
-from .migthy_types import MIGHTYENV, TypeKwargs, retrieve_class
+from .mighty_types import MIGHTYENV, DACENV, CARLENV, TypeKwargs, retrieve_class
 from .update_utils import polyak_update
 
 __all__ = [
     "MIGHTYENV",
+    "DACENV",
+    "CARLENV",
     "make_mighty_env",
     "TypeKwargs",
     "retrieve_class",
