+++ mighty/mighty_utils/mighty_types.py
